s/drag_coefficient/ballistic_coefficient/
Uh-oh. GNU Ballistics, the underlying C lib, uses the wrong term for the
ballistic coefficient.  The BC is what goes into a ballistics calculator.
The drag coefficient is part of the underlying physics but is not generally
an input.  They are different entities.  GNU Ballistics is dealing with a
ballistic coefficient but calls it a drag coefficient.  This is unfortunate.

This is another reason to consider basing this Ruby lib on a different
underlying ballistics lib.  JBM Ballistics, probably.